% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filterICPSRFiles}
\alias{filterICPSRFiles}
\title{Filter the (very large) ICPSR data files to include only the records for a single state}
\usage{
filterICPSRFiles(filePaths, state, outputDir)
}
\arguments{
\item{filePaths}{paths to the files to filter}

\item{state}{two-letter state code for records to select}

\item{outputDir}{directory into which filtered files are written}
}
\value{
a named character vector mapping input file paths to output (filtered) file paths
}
\description{
Filter the (very large) ICPSR data files to include only the records for a single state
}
