% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{loadICPSR}
\alias{loadICPSR}
\title{Load NIBRS records from ICPSR input files}
\usage{
loadICPSR(conn = DBI::dbConnect(RMySQL::MySQL(), host = "localhost", dbname =
  "nibrs_analytics", username = "root"), agencyFile, incidentFile, arresteeFile,
  incidentRecords = -1, arresteeRecords = -1)
}
\arguments{
\item{conn}{the database connection to use}

\item{agencyFile}{path to the ICPSR agency file}

\item{incidentFile}{path to the ICPSR NIBRS incident-level file}

\item{arresteeFile}{path to the ICPSR NIBRS arrestee-level file}

\item{incidentRecords}{number of records to read from the incident file, or -1 to read all records}

\item{arresteeRecords}{number of records to read from the arrestee file, or -1 to read all records}
}
\value{
a list of tibbles, where the name of each list member is the name of the table in the database
}
\description{
Load NIBRS records from the ICPSR input files, write them to the staging database (structured per the model provided in this package's same
GitHub repo), and return the loaded tables in a list of tibbles.
}
